react-gui/tests/requirements.txt
# Test requirements for React GUI foundation
# Comprehensive testing dependencies with defensive programming support

# Core testing framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0
pytest-json-report>=1.5.0
pytest-html>=3.2.0
pytest-xdist>=3.3.0  # Parallel test execution

# WebSocket testing
websockets>=11.0.0
pytest-websocket>=0.1.0

# HTTP testing
requests>=2.31.0
httpx>=0.24.0
pytest-httpx>=0.21.0

# Async testing utilities
asyncio-mqtt>=0.13.0
aioresponses>=0.7.4

# Mock and fixture utilities
responses>=0.23.0
factory-boy>=3.3.0
faker>=19.0.0

# Data validation testing
pydantic>=2.0.0
jsonschema>=4.19.0

# Performance testing
pytest-benchmark>=4.0.0
memory-profiler>=0.61.0

# Security testing
bandit>=1.7.5
safety>=2.3.0

# Code quality
flake8>=6.0.0
black>=23.7.0
isort>=5.12.0
mypy>=1.5.0

# Coverage reporting
coverage>=7.3.0
coverage-badge>=1.1.0

# Test data generation
hypothesis>=6.82.0
mimesis>=10.1.0

# Database testing (if needed)
pytest-postgresql>=5.0.0
pytest-redis>=3.0.0

# Frontend testing dependencies (for integration)
selenium>=4.11.0
playwright>=1.37.0
pytest-playwright>=0.4.0

# API testing
tavern>=2.3.0
schemathesis>=3.19.0

# Load testing
locust>=2.16.0

# Reporting and documentation
allure-pytest>=2.13.0
pytest-clarity>=1.0.1
pytest-sugar>=0.9.7

# Development utilities
ipdb>=0.13.13
pytest-watch>=4.2.0
pytest-testmon>=2.0.0

# Type checking for tests
types-requests>=2.31.0
types-mock>=5.1.0

# Configuration management
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Logging and debugging
structlog>=23.1.0
colorlog>=6.7.0

# Time and date utilities
freezegun>=1.2.2
time-machine>=2.12.0

# File system testing
pyfakefs>=5.2.4

# Network testing
pytest-socket>=0.6.0
